+++ feature/review/src/commonTest/kotlin/com/hulkdx/findprofessional/feature/review/ReviewUseCaseTest.kt
@@ -98,7 +98,7 @@ class ReviewUseCaseTest {
         var findAllReviewsPage = -1000
 
         // @formatter:off
-        override suspend fun findAll(professionalId: Int, page: Int, pageSize: Int):List<ProfessionalReview>  {
+        override suspend fun findAll(professionalId: Long, page: Int, pageSize: Int):List<ProfessionalReview>  {
             findAllReviewsCalled++
             findAllReviewsPage = page
             return response()
